+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/aware/IntelliJPlatformAware.kt
@@ -1,0 +1,8 @@
+// Copyright 2000-2026 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+
+package org.jetbrains.intellij.platform.gradle.tasks.aware
+
+/**
+ * Marker interface for tasks that require IntelliJ Platform Gradle Plugin task preconfiguration.
+ */
+interface IntelliJPlatformAware